two warnings gone, still one to go

+++ nassl/_nassl/nassl_SSL_CTX.c
@@ -191,7 +191,7 @@ static int pem_passwd_cb(char *buf, int size, int rwflag, void *userdata) {
 
     // NUL-terminated string as it will come from Python
     passwordSize = strlen(passwordBuf) + 1;
-    if (passwordSize > size){  // Not enough space in OpenSSL's buffer
+    if ((int)passwordSize > size){  // Not enough space in OpenSSL's buffer
         return 0;
     }
